Move the json objs to own file
